+++ app/Http/Controllers/CharacterController.php
@@ -2,8 +2,8 @@
 
 namespace App\Http\Controllers;
 
+use HaydenPierce\ClassFinder\ClassFinder;
 use Illuminate\Http\Request;
-use App\Characters;
 
 class CharacterController extends Controller
 {
@@ -14,8 +14,15 @@ class CharacterController extends Controller
      */
     public function index()
     {
-        $characters = Characters::all();
-        //dd($characters);
+        $characterClasses = ClassFinder::getClassesInNamespace('App\Characters');
+        $characters = [];
+
+        foreach ($characterClasses as $character) {
+            $a = new $character;
+
+            array_push($characters, $a);
+        }
+
         return view('characters', compact('characters'));
     }
 }
