+++ public/css/style.css
@@ -183,7 +183,7 @@ body {
 	position: absolute;
 	top: 0;
 	left: 0;
-	width: 100%
+	width: 100%;
 	height: 100%;
 }
